Move account related functions from Pica to new PicaOld class and create PicaLBS API
